pyproject.toml
[project]
name = "trackapp-env"
version = "0.1.0"
description = "RL environment for track & field Django app bug-fixing tasks"
requires-python = ">=3.11"
dependencies = [ "hud-python[agents]>=0.5.17", "click>=8.0.0", "mcp[cli]>=1.10.1", "numpy>=1.26.0", "packaging>=21.0", "pillow", "pydantic>=2.11.4",]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [ "pytest>=8.0.0", "ruff==0.11.2", "pre-commit>=3.5.0",]

[project.scripts]
hud_eval = "cli:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.hud]
image = "trackapp-rl-env:dev"

[tool.ruff.lint]
select = [ "E", "F", "B", "I", "C4", "UP",]
ignore = [ "E501",]
fixable = [ "ALL",]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.hatch.build.targets.wheel]
packages = [ ".",]
